NEW: Created BaseITST class with common env creation
