+++ api/src/test/java/io/strimzi/api/kafka/model/KafkaMirrorMakerCrdIT.java
@@ -4,13 +4,13 @@
  */
 package io.strimzi.api.kafka.model;
 
-import io.strimzi.test.annotations.Namespace;
-import io.strimzi.test.annotations.Resources;
-import io.strimzi.test.extensions.StrimziExtension;
-import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.KubeClusterException;
 import org.junit.jupiter.api.Test;
+
+import java.util.Collections;
 
 import static org.junit.Assert.assertTrue;
 
@@ -20,9 +20,6 @@ import static org.junit.Assert.assertTrue;
  * I.e. that such instance resources obtained from POJOs are valid according to the schema
  * validation done by K8S.
  */
-@ExtendWith(StrimziExtension.class)
-@Namespace(KafkaMirrorMakerCrdIT.NAMESPACE)
-@Resources(value = TestUtils.CRD_KAFKA_MIRROR_MAKER, asAdmin = true)
 public class KafkaMirrorMakerCrdIT extends AbstractCrdIT {
 
     public static final String NAMESPACE = "kafkamirrormaker-crd-it";
@@ -79,8 +76,21 @@ public class KafkaMirrorMakerCrdIT extends AbstractCrdIT {
     }
 
     @Test
-    public void testKafkaMirrorMakerWithTemplate() {
+    void testKafkaMirrorMakerWithTemplate() {
         createDelete(KafkaMirrorMaker.class, "KafkaMirrorMaker-with-template.yaml");
+    }
+
+    @BeforeAll
+    void setupEnvironment() {
+        createNamespaces(NAMESPACE);
+        createCustomResources(Collections.singletonList(
+                TestUtils.CRD_KAFKA_MIRROR_MAKER));
+    }
+
+    @AfterAll
+    void teardownEnvironment() {
+        deleteCustomResources();
+        deleteNamespaces();
     }
 }
