+++ test/src/main/java/io/strimzi/test/annotations/EnvVariables.java
@@ -24,6 +24,7 @@ import java.lang.annotation.Target;
  * }
  * </pre>
  */
+@Deprecated
 @Target({})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface EnvVariables {
